Did a bit of reafactoring. Moved HomeView inside of the Home folder and combined it with HomeSearchView, since they are redundant. It makes more sense to extract the individual texts out into like a "SongResultView" and the search bar as a "SearchView", but not so much wrap the HomeView inside of another empty view.
